refactor(migrate-history): sealed ClaimAttempt + heartbeat-by-time + force-reset TOCTOU
Acts on the latest review round + my own pushback on the @Suppress count.

P1 — force-reset TOCTOU between Guard 1 (in-memory) and Guard 2 (DB row)
  Window between current() check and findById was microseconds wide but
  matched exactly the gap between an external startAsync's compareAndSet
  and its preflight tryInsert — a freshly-claimed live import could slip
  past both guards. Re-check current() after the row read; if RUNNING,
  treat as Guard 1 hit (409). Cheap defensive double-check.

P1 — heartbeat by wall-clock, not commit count
  Was every 50 commits, which silently failed two corner cases:
  (1) chains < 50 commits — heartbeat never fires;
  (2) pathologically slow per-commit step (Groovy DSL parse on a huge
      tree) — 50 commits could exceed the 30-min staleness threshold.
  Now: 5-min interval based on Instant.now(). Both edge cases gone;
  cadence comfortably under STALE_IN_PROGRESS_THRESHOLD; no extra writes
  on small chains.

Refactor — sealed ClaimAttempt removes LoopWithTooManyJumpStatements suppress
  startAsync was a single while-loop with 5 different jump statements
  (return × 2, continue × 2, throw × 1) plus a nested catch path. The
  @Suppress comment claimed "the loop IS the pattern" but it was actually
  obscuring the load-bearing ordering. Decomposed into:
    - sealed interface ClaimAttempt with 4 cases (Attached, CrossKindConflict,
      Retry, Claimed)
    - private tryClaim() returning a ClaimAttempt
    - private submitToExecutor() handling the rejection-rollback path
    - startAsync() reduced to a clean while-loop with `when` over the
      sealed cases — exactly one terminal action per case.
  Same behavior, far easier to follow, no @Suppress needed.
  Mirrored in both MigrationJobServiceImpl and HistoryMigrationJobServiceImpl.

Refactor — extract StubGitHistoryImportStateRepository to its own file
  Test file had @file:Suppress("MaxLineLength") for boilerplate JpaRepository
  overrides — file-level suppression hid future drift in real test code.
  Extracted the stub to StubGitHistoryImportStateRepository.kt with scoped
  suppression on JUST that boilerplate file. Test file gets default 140-char
  limit back so genuine long-line drift would now fail CI again.

@Suppress count: 13 → 9 (kept those that are genuinely unavoidable: 1×
TooManyFunctions on AdminControllerV4 [matches ComponentControllerV4],
2× TooGenericExceptionCaught on submitToExecutor [Throwable so executor
rejection ALWAYS unwinds slot], 2× UnusedPrivateProperty on @MockBean
[Spring needs the bean present for context wiring], 3× DEPRECATION on
JpaRepository overrides [Spring still declares them], 1× MaxLineLength
on the extracted stub [scope-limited]).
